Code clean up

+++ app/src/main/java/com/example/footixappbachelorarbeit/viewModelLiveData/Session.kt
@@ -8,7 +8,7 @@ import androidx.room.PrimaryKey
 data class Session(
     @PrimaryKey(autoGenerate = true) val id: Int?,
     @ColumnInfo(name = "current_date") var currentDate: String?,
-    @ColumnInfo(name = "total_distance") var totalDistance: Float?,
+    @ColumnInfo(name = "total_distance") var totalDistance: Double?,
     @ColumnInfo(name = "max_speed") var maxSpeed: Float?,
     @ColumnInfo(name = "run_time") var runTime: String?
 )
